# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Garage API"
  description: "API for managing an auto garage, including customers, inventory, mechanics, and service tickets."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
definitions:
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: "Resource not found."
  ValidationErrorResponse:
    type: object
    properties:
      messages:
        type: object
        additionalProperties:
          type: array
          items:
            type: string
        example:
          email: ["Invalid email format.", "Email already exists."]
          password: ["Field may not be null."]
  MessageResponse:
    type: object
    properties:
      message:
        type: string
        example: "Operation successful."

  # Customer Definitions
...
  MechanicCore:
    type: object
    properties:
      name:
        type: string
        example: "Jane Smith"
      email:
        type: string
        format: email
        example: "jane.smith@example.com"
      phone:
        type: string
        example: "555-5678"
      salary:
        type: string # Representing decimal as string
        example: "60000.00"
    required:
      - name
      - email
      - salary
  MechanicPayload:
    allOf:
      - $ref: "#/definitions/MechanicCore"
      - type: object
        properties:
          password:
            type: string
            format: password
            example: "anotherSecurePassword"
            writeOnly: true
        required:
          - password
  MechanicResponse:
    allOf:
      - $ref: "#/definitions/MechanicCore"
      - type: object
        properties:
          id:
            type: integer
            format: int64
            readOnly: true
            example: 1
          service_tickets:
            type: array
            items:
              $ref: "#/definitions/ServiceTicketSummary"
            readOnly: true
  MechanicLoginPayload:
    type: object
    properties:
      email:
        type: string
        format: email
        example: "jane.smith@example.com"
      password:
        type: string
        format: password
        example: "anotherSecurePassword"
    required:
      - email
      - password

  # Inventory Definitions
...
  ServiceTicketSummary: # For nesting in Customer/Mechanic
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      vin:
        type: string
        example: "123ABC456DEF789GHI"
      service_date:
        type: string
        format: date
        example: "2024-07-26"
      service_desc:
        type: string
        example: "Oil change and tire rotation"
...
  AuthTokenResponse:
    type: object
    properties:
      status:
        type: string
        example: "success"
      message:
        type: string
        example: "Successfully Logged In"
      auth_token:
        type: string
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
paths:
...
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: "Login mechanic"
      description: "Authenticates a mechanic and returns an auth token."
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/MechanicLoginPayload"
          examples:
            application/json:
              email: "jane.smith@example.com"
              password: "anotherSecurePassword"
      responses:
        "200":
          description: "Login successful"
          schema:
            $ref: "#/definitions/AuthTokenResponse"
          examples:
            application/json:
              status: "success"
              message: "Successfully Logged In"
              auth_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJtZWNoYW5pYy0xIiwibmFtZSI6IkphbmUgU21pdGgiLCJpYXQiOjE1MTYyMzkwMjJ9.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        "400":
          description: "Invalid payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Invalid payload, expecting username and password" # Route message
        "401":
          description: "Invalid email or password"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Invalid email or password"
  /mechanics/:
    post:
      tags:
        - Mechanics
      summary: "Create a new mechanic"
      # security: - BearerAuth: [] # Route has @mec_token_required commented out
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/MechanicPayload"
          examples:
            application/json:
              name: "Mechanic Bob"
              email: "bob.builder@example.com"
              phone: "555-1122"
              salary: "65000.00"
              password: "fixitFelix"
      responses:
        "201":
          description: "Mechanic created successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              id: 2
              name: "Mechanic Bob"
              email: "bob.builder@example.com"
              phone: "555-1122"
              salary: "65000.00"
              service_tickets: []
        "400":
          description: "Validation error or email already exists"
          schema:
            oneOf:
              - $ref: "#/definitions/ErrorResponse"
              - $ref: "#/definitions/ValidationErrorResponse"
          examples:
            application/json_exists:
              error: "Email already associated with a mechanic."
            application/json_validation:
              messages:
                salary: ["Not a valid number."]
    get:
      tags:
        - Mechanics
      summary: "Get all mechanics"
      responses:
        "200":
          description: "A list of mechanics"
          schema:
            type: array
            items:
              $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              - id: 1
                name: "Jane Smith"
                email: "jane.smith@example.com"
                phone: "555-5678"
                salary: "60000.00"
                service_tickets: []
              - id: 2
                name: "Mechanic Bob"
                email: "bob.builder@example.com"
                phone: "555-1122"
                salary: "65000.00"
                service_tickets: []
...
  /mechanics/{mechanic_id}:
    put:
      tags:
        - Mechanics
      summary: "Update mechanic by ID"
      # security: - BearerAuth: [] # Route has @mec_token_required commented out
      parameters:
        - name: mechanic_id
          in: path
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/MechanicPayload"
          examples:
            application/json:
              name: "Jane A. Smith"
              email: "jane.a.smith@example.com"
              phone: "555-5679"
              salary: "62000.00"
              password: "newSecurePasswordForJane"
      responses:
        "200":
          description: "Mechanic updated successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              id: 1
              name: "Jane A. Smith"
              email: "jane.a.smith@example.com"
              phone: "555-5679"
              salary: "62000.00"
              service_tickets: []
        "400":
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
          examples:
            application/json:
              messages:
                email: ["Not a valid email address."]
        "404":
          description: "Mechanic not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Mechanic not found."
    delete:
      tags:
        - Mechanics
      summary: "Delete mechanic by ID"
      security:
        - BearerAuth: [] # @mec_token_required is active
      parameters:
        - name: mechanic_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: "Mechanic deleted successfully"
          schema:
            $ref: "#/definitions/MessageResponse"
          examples:
            application/json:
              message: "Mechanic id: 2 successfully deleted by user mechanic-1."
        "400":
          description: "Invalid mechanic ID format"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Invalid mechanic ID format in URL."
        "403":
          description: "Cannot delete the default mechanic"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Cannot delete the default mechanic account."
        "404":
          description: "Mechanic not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Mechanic not found."
        "500":
          description: "System configuration error (default mechanic not found)"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "System configuration error: Default customer not found. Please contact support." # Note: Error message in route refers to 'customer'
...